feat: 서비스에서 개인/전체 카테고리 조회 로직 분기
