+++ backend/src/main/java/com/staccato/category/controller/CategoryController.java
@@ -2,8 +2,10 @@ package com.staccato.category.controller;
 
 import java.net.URI;
 import java.time.LocalDate;
+
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Min;
+
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -16,6 +18,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+
 import com.staccato.category.controller.docs.CategoryControllerDocs;
 import com.staccato.category.domain.Scope;
 import com.staccato.category.service.CategoryService;
@@ -33,6 +36,7 @@ import com.staccato.category.service.dto.response.CategoryStaccatoLocationRespon
 import com.staccato.config.auth.LoginMember;
 import com.staccato.config.log.annotation.Trace;
 import com.staccato.member.domain.Member;
+
 import lombok.RequiredArgsConstructor;
 
 @Trace
@@ -68,7 +72,7 @@ public class CategoryController implements CategoryControllerDocs {
             @RequestParam(value = "specificDate") LocalDate specificDate,
             @RequestParam(value = "scope") Scope scope
     ) {
-        CategoryNameResponses categoryNameResponses = categoryService.readAllCategoriesByDateAndIsShared(member, specificDate, false, scope);
+        CategoryNameResponses categoryNameResponses = categoryService.readAllCategoriesByDateAndIsShared(member, specificDate, scope);
         return ResponseEntity.ok(categoryNameResponses);
     }
