+++ backend/src/main/java/com/staccato/category/domain/Scope.java
@@ -3,4 +3,12 @@ package com.staccato.category.domain;
 public enum Scope {
     ALL,
     PRIVATE;
+
+    public boolean isAll() {
+        return this == ALL;
+    }
+
+    public boolean isPrivate() {
+        return this == PRIVATE;
+    }
 }
